Rename storybook -> storybook_rails
