+++ lib/generators/storybook_rails/templates/stories.html.erb.tt
