+++ lib/generators/storybook_rails/templates/stories.rb.tt
